Refactor: Introduce smarterwebsearchmcp package
I moved core application logic into a new 'smarterwebsearchmcp' package.
This includes:
- manager.py
- planner_agent.py
- search_agent.py
- search_critic.py
- web_search_tool.py
- printer.py

I updated all import statements in the affected files (`main.py` and files within the new package) to reflect the new module paths.

I modified `pyproject.toml` to remove the explicit `py-modules` list and replaced it with `packages = ["smarterwebsearchmcp"]` to correctly build and distribute the new package.

The OpenAI SDK dependency (`openai>=1.0.0`) was already present in `pyproject.toml` and remains unchanged. The placeholder OpenAI SDK import paths in `smarterwebsearchmcp/manager.py` have been kept as they were, as they passed `python -m compileall` and further verification would require runtime checks.
