+++ smarterwebsearchmcp/printer.py
