+++ smarterwebsearchmcp/search_critic.py
